+++ config/application.rb
@@ -6,8 +6,8 @@ require 'rails/all'
 
 Bundler.require(*Rails.groups)
 
-module MyLogSays
-  # Rails application configuration for MyLogSays.
+module LogLady
+  # Rails application configuration for LogLady.
   class Application < Rails::Application
     config.load_defaults 8.1